requirements/requirements.txt
####### requirements.txt #######

asyncio
discord
python-dotenv
